Fixed Issue Due to Apollo Caching Storing Wrong Data
The attributes property has an ID feild which Apollo uses to cache its value the id with value "size" is duplicated with the shoe and the jacket products and the shoe gets fetched first and its "Size" values gets cached and skipped with the jacket since they have the same ID, the same happens for the color attribute.

The issue is bandaided by skipping the __typename which is need alongside the ID for apollo to cache the value. More info :
https://stackoverflow.com/questions/54613216/apollo-query-has-the-wrong-data-from-cache

And This article:

https://kamranicus.com/graphql-apollo-object-caching/